Move compiler to separate file
